gar: Enhance values for SPKG_SOURCEURL (for VENDOR in pkginfo)
SPKG_SOURCEURL is now set to the first of VENDOR_URL, MASTER_SITES or
GIT_REPOS.  It was originally only MASTER_SITES, but this doesn't work
when GIT_REPOS but no MASTER_SITES are set.  GIT_REPOS on it's own
isn't a good stand-in as it's a less 'pretty' url.  Allowing an
explicit VENDOR_URL gives control to the maintainer, retains backward
compatibility by default and extends the old behaviour with a new
fallback possibility.



git-svn-id: https://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@7897 d3b55034-1cff-0310-a425-aefe953e1e90