gar/v2: Transitional change for var renaming GAR{NAME,VERSION} -> {NAME,VERSION}
As a preparation for the introduction of GARTYPE, the two variables GARNAME
and GARVERSION will be renamed to NAME and VERSION (to better reflect their
relevance to the package, not GAR). This change renames these variables
within GAR while - for now - leaving existing build descriptions functional.

After adjusting our documentation and sending a heads up to maintainers@,
the next step will then obsolete the old variables.



git-svn-id: https://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@11860 d3b55034-1cff-0310-a425-aefe953e1e90