checkpkg: Dependency checking engine update
The new function NeedFile had some side effects which I didn't initially see.
Symptoms included surplus dependency reporting, when there in fact weren't
any.  The problem was that the library checking code was using the old way of
reporting dependencies.  You can really have only one place in the processing
workflow in which you can decide which dependencies report as required and
which as surplus.  Thich change removes the old way of reporting errors from
the dependency checks.

A new API function is introduced: NeedPackage.  It works the same way that
NeedFile does, but it reports a package name rather than a file name.

As a side note, I've simplified some unit tests by using comparators.  Instead
of specifying each argument of an expected function call, I only say that
there should be e.g. 3 arguments of certain types.

It's a big patch, but I didn't see a good way of chopping it into smaller
ones without committing incomplete or broken code.

I feel that the part of code I've been working with, has to be rewritten,
becaus it grew organically, and its logic is not clear from looking at the
code.

git-svn-id: https://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@12068 d3b55034-1cff-0310-a425-aefe953e1e90