csw-upload-pkg: A more exhaustive unit test
A new, data-driven unit test, which allows to define conditions in a very
concise way, and expect the right output.

It has a problem that it stops on the first failed tests and doesn't do any
other tests.  It could be rewritten using metaclases, but they would make
the code more complicated than it currently is, which is quite complicated
already.

git-svn-id: https://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@13347 d3b55034-1cff-0310-a425-aefe953e1e90