csw-upload-pkg: Sort filenames according to OS rel
When a bulk of files is uploaded, e.g. SunOS5.9 and SunOS5.10 files are
uploaded together, SunOS5.9 files are lexicographically last.  In
certain operation it's better to have them sorted according to the OS
release.  The typical usage scenario is when there are only 5.9 packages
in the database, and we're uploading a mix of 5.9 and 5.10 packages.
Without this sorting, only 5.9 would be in the catalog in result.  The
5.10 packages would be initially inserted into the database, but then
they would be overwritten by 5.9 packages.

git-svn-id: https://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@14117 d3b55034-1cff-0310-a425-aefe953e1e90