checkpkg: Relocatable package support
GAR packages are not relocatable, but there is a couple of relocatable
packages in the catalog.

There's one function responsible from returning the basedir, but without the
trailing slash, because that't the convention of keeping paths in checkpkg.

When constructing pkgmap however, the slash needs to be added back - paths in
the files table and in the pkgmap list need to be absolute.

git-svn-id: https://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@14921 d3b55034-1cff-0310-a425-aefe953e1e90