pkgdb: When hachoir fails, display the exception
When indexing the whole filesystem, hachoir_parser fails on a number of
files. For now, we're catching the exception and moving on. It might
bite us later, but I have no good idea how to deal with it right now.

This change makes the exception message be printed to stdout.

git-svn-id: https://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@17553 d3b55034-1cff-0310-a425-aefe953e1e90