pkgdb: More progress bar during system pkg import
After creating the fake package objects, it's necessary do create CswFile
objects related to these packages so that the buildfarm database is able to
map from files to packages. It can take a couple minutes, so let's display
a progress bar for it as well.

Before the CswFile registration, display a msg, so people have a better idea
what's the progress bar for.

git-svn-id: https://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@19711 d3b55034-1cff-0310-a425-aefe953e1e90