checkpkg: When parsing ldd output fails, log it
It's useful to have the complete output from ldd logged. More so, if you're
running in screen with a long scrollback buffer which gives you an option to
save that output into a file on disk.

We could potentially save the ldd output to disk instead of printing to
screen. Suggestions are welcome.

git-svn-id: https://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@20106 d3b55034-1cff-0310-a425-aefe953e1e90