checkpkg: Free magic cookies on deletion
I was hoping this was the memory leak, but it looks like it isn't it. I still
suspect something around libmagic though, mainly because the process RSS size
is way bigger than reported object sizes in Python memory profiler. libmagic is
the only C module that is used, and this symptom is likely produced by
a C module.

git-svn-id: https://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@20799 d3b55034-1cff-0310-a425-aefe953e1e90