csw-upload-pkg: Error msg when pkg unknown
When a package is not in the DB, csw-upload-pkg will not work on the login
host. This is the case for many reasons, including:

- login is a sparc host, and we currently are not able to index intel packages
	on a sparc box
- something's messed up with path searches, so the installed csw-upload-pkg
	utility is unable to find the pkgtrans helper
- not having an indexed package at this point is unusual, so this scenario is
	prone to breaking silently

Therefore, it's better to not even try to index the package, but display an
error message instead, so maintainers have a better idea what to do.

git-svn-id: https://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@20871 d3b55034-1cff-0310-a425-aefe953e1e90