pkgdb: Faster catalog generation
Instead of lots of back-and-forth over the REST interface, we now have
a single endpoint which gives us complete catalog information. All we need to
do is to write that to a text file and call it a catalog.

The server side has to do a little more work, for example filter out all
non-CSW packages from dependencies.

git-svn-id: https://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@21086 d3b55034-1cff-0310-a425-aefe953e1e90