add support for per-os-release allowed library versions dependencies
The check for usage of forbidden library version interface must take
into account the os release to accomodate differences between releases.
For example it's fine to only allowed libc version < SUNW_1.22.6 under
Solaris 10, but it doesn't make sense under Solaris 11 as SUNW_1.23
was provided from the very first release.

This patch also include a new comparison function to be able to compare
version dependencies so that we do not have to declare all versions
(SUNW_0.8, SUNW_0.9...) but only the maximal version allowed.

git-svn-id: https://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@21904 d3b55034-1cff-0310-a425-aefe953e1e90